refactor DaqScalers
