+++ common-tools/clas-analysis/src/main/java/org/jlab/analysis/eventmerger/FilterTrigger.java
@@ -1,5 +1,5 @@
 package org.jlab.analysis.eventmerger;
-import org.jlab.detector.decode.DaqScalersSequence;
+import org.jlab.detector.scalers.DaqScalersSequence;
 import org.jlab.jnp.hipo4.data.*;
 import org.jlab.jnp.hipo4.io.HipoReader;
 import org.jlab.jnp.utils.data.*;
